Simplify hazards state
