--- src/state/data-selection/hazards/hazard-selection.ts
@@ -1,18 +1,0 @@
-import _ from 'lodash';
-import { atomFamily, RecoilValue } from 'recoil';
-
-export const hazardSelectionState = atomFamily({
-  key: 'hazardSelectionState',
-  default: false,
-});
-
-interface TransactionGetterInterface {
-  get<T>(a: RecoilValue<T>): T;
-}
-
-export function getHazardSelectionAggregate(
-  { get }: TransactionGetterInterface,
-  hazards: readonly string[],
-) {
-  return _.fromPairs(hazards.map((group) => [group, get(hazardSelectionState(group))]));
-}
