+++ src/state/layers/data-layers/hazards.ts
@@ -5,7 +5,7 @@ import { truthyKeys } from '@/lib/helpers';
 
 import { hazardViewLayer } from '@/config/hazards/hazard-view-layer';
 import { dataParamsByGroupState } from '@/state/data-params';
-import { hazardVisibilityState } from '@/state/data-selection/hazards/hazard-visibility';
+import { hazardVisibilityState } from '@/state/data-selection/hazards';
 
 export const hazardLayerState = selector<ViewLayer[]>({
   key: 'hazardLayerState',
